ìexit 요청 유저만 하도록 처리

+++ back/B310_Back/src/main/java/com/ssafy/B310/service/ParticipationServiceImpl.java
@@ -68,7 +68,9 @@ public class ParticipationServiceImpl implements ParticipationService {
 
         List<Participation> joinedUser = participationRepository.findByRoom(room);
         for (Participation participation : joinedUser)
-            participationRepository.delete(participation);
+            if (participation.getUser().getUserNum() == userNum) {
+                participationRepository.delete(participation);
+            }
         return 1;
     }
